[gcode,grbl parsers] expose tests to rawkit

+++ lib/gcode-parser/test/gcode-parser-test.cpp
@@ -1,11 +1,11 @@
-#include <doctest/doctest.h>
+#include <doctest.h>
 
 #define ok CHECK
 char debug_buf[4096] = "\0";
 #define gcode_debug(_f, ...) \
   do { sprintf((debug_buf),_f,##__VA_ARGS__); INFO(debug_buf); } while(false)
 
-#include "include/rawkit/gcode/parser.h"
+#include "../include/rawkit/gcode/parser.h"
 
 TEST_CASE("[gcode] letter letter fails") {
   GCODEParser p;
@@ -214,75 +214,75 @@ TEST_CASE("[grbl,gcode] commands") {
   ok(p.line(3)->type == GCODE_LINE_TYPE_COMMAND_FEED_HOLD);
   ok(p.line_count() == 5);
 
-  ok(p.push(0x84) == GCODE_RESULT_TRUE);
+  ok(p.push((uint8_t)0x84) == GCODE_RESULT_TRUE);
   ok(p.line(4)->type == GCODE_LINE_TYPE_COMMAND_SAFETY_DOOR);
   ok(p.line_count() == 6);
 
-  ok(p.push(0x85) == GCODE_RESULT_TRUE);
+  ok(p.push((uint8_t)0x85) == GCODE_RESULT_TRUE);
   ok(p.line(5)->type == GCODE_LINE_TYPE_COMMAND_JOG_CANCEL);
   ok(p.line_count() == 7);
 
-  ok(p.push(0x90) == GCODE_RESULT_TRUE);
+  ok(p.push((uint8_t)0x90) == GCODE_RESULT_TRUE);
   ok(p.line(6)->type == GCODE_LINE_TYPE_COMMAND_FEED_OVERRIDE_SET_100_PERCENT);
   ok(p.line_count() == 8);
 
-  ok(p.push(0x91) == GCODE_RESULT_TRUE);
+  ok(p.push((uint8_t)0x91) == GCODE_RESULT_TRUE);
   ok(p.line(7)->type == GCODE_LINE_TYPE_COMMAND_FEED_OVERRIDE_INCREASE_10_PERCENT);
   ok(p.line_count() == 9);
 
-  ok(p.push(0x92) == GCODE_RESULT_TRUE);
+  ok(p.push((uint8_t)0x92) == GCODE_RESULT_TRUE);
   ok(p.line(8)->type == GCODE_LINE_TYPE_COMMAND_FEED_OVERRIDE_DECREASE_10_PERCENT);
   ok(p.line_count() == 10);
 
-  ok(p.push(0x93) == GCODE_RESULT_TRUE);
+  ok(p.push((uint8_t)0x93) == GCODE_RESULT_TRUE);
   ok(p.line(9)->type == GCODE_LINE_TYPE_COMMAND_FEED_OVERRIDE_INCREASE_1_PERCENT);
   ok(p.line_count() == 11);
 
-  ok(p.push(0x94) == GCODE_RESULT_TRUE);
+  ok(p.push((uint8_t)0x94) == GCODE_RESULT_TRUE);
   ok(p.line(10)->type == GCODE_LINE_TYPE_COMMAND_FEED_OVERRIDE_DECREASE_1_PERCENT);
   ok(p.line_count() == 12);
 
-  ok(p.push(0x95) == GCODE_RESULT_TRUE);
+  ok(p.push((uint8_t)0x95) == GCODE_RESULT_TRUE);
   ok(p.line(11)->type == GCODE_LINE_TYPE_COMMAND_SEEK_OVERRIDE_SET_100_PERCENT);
   ok(p.line_count() == 13);
 
-  ok(p.push(0x96) == GCODE_RESULT_TRUE);
+  ok(p.push((uint8_t)0x96) == GCODE_RESULT_TRUE);
   ok(p.line(12)->type == GCODE_LINE_TYPE_COMMAND_SEEK_OVERRIDE_SET_50_PERCENT);
   ok(p.line_count() == 14);
 
-  ok(p.push(0x97) == GCODE_RESULT_TRUE);
+  ok(p.push((uint8_t)0x97) == GCODE_RESULT_TRUE);
   ok(p.line(13)->type == GCODE_LINE_TYPE_COMMAND_SEEK_OVERRIDE_SET_25_PERCENT);
   ok(p.line_count() == 15);
 
-  ok(p.push(0x99) == GCODE_RESULT_TRUE);
+  ok(p.push((uint8_t)0x99) == GCODE_RESULT_TRUE);
   ok(p.line(14)->type == GCODE_LINE_TYPE_COMMAND_SPINDLE_OVERRIDE_SET_100_PERCENT);
   ok(p.line_count() == 16);
 
-  ok(p.push(0x9A) == GCODE_RESULT_TRUE);
+  ok(p.push((uint8_t)0x9A) == GCODE_RESULT_TRUE);
   ok(p.line(15)->type == GCODE_LINE_TYPE_COMMAND_SPINDLE_OVERRIDE_INCREASE_10_PERCENT);
   ok(p.line_count() == 17);
 
-  ok(p.push(0x9B) == GCODE_RESULT_TRUE);
+  ok(p.push((uint8_t)0x9B) == GCODE_RESULT_TRUE);
   ok(p.line(16)->type == GCODE_LINE_TYPE_COMMAND_SPINDLE_OVERRIDE_DECREASE_10_PERCENT);
   ok(p.line_count() == 18);
 
-  ok(p.push(0x9C) == GCODE_RESULT_TRUE);
+  ok(p.push((uint8_t)0x9C) == GCODE_RESULT_TRUE);
   ok(p.line(17)->type == GCODE_LINE_TYPE_COMMAND_SPINDLE_OVERRIDE_INCREASE_1_PERCENT);
   ok(p.line_count() == 19);
 
-  ok(p.push(0x9D) == GCODE_RESULT_TRUE);
+  ok(p.push((uint8_t)0x9D) == GCODE_RESULT_TRUE);
   ok(p.line(18)->type == GCODE_LINE_TYPE_COMMAND_SPINDLE_OVERRIDE_DECREASE_1_PERCENT);
   ok(p.line_count() == 20);
 
-  ok(p.push(0x9E) == GCODE_RESULT_TRUE);
+  ok(p.push((uint8_t)0x9E) == GCODE_RESULT_TRUE);
   ok(p.line(19)->type == GCODE_LINE_TYPE_COMMAND_TOGGLE_SPINDLE_STOP);
   ok(p.line_count() == 21);
 
-  ok(p.push(0xA0) == GCODE_RESULT_TRUE);
+  ok(p.push((uint8_t)0xA0) == GCODE_RESULT_TRUE);
   ok(p.line(20)->type == GCODE_LINE_TYPE_COMMAND_TOGGLE_FLOOD_COOLANT);
   ok(p.line_count() == 22);
 
-  ok(p.push(0xA1) == GCODE_RESULT_TRUE);
+  ok(p.push((uint8_t)0xA1) == GCODE_RESULT_TRUE);
   ok(p.line(21)->type == GCODE_LINE_TYPE_COMMAND_TOGGLE_MIST_COOLANT);
   ok(p.line_count() == 23);
 }
